Split readHeader into its own module + dedicated tests
Move the RowBinaryWithNamesAndTypes header read out of compile.ts into
src/header.ts (with a RowBinaryHeader return type), and extract the shared
test encoder into tests/rowBinaryWriter.ts so the header tests can live in
tests/readHeader.test.ts alongside the compiler tests.
