+++ skills/clickhouse-js-node-rowbinary-parser/src/compile.ts
@@ -28,7 +28,7 @@ import {
 } from "@clickhouse/datatype-parser";
 
 import type { Reader, Cursor } from "./core.js";
-import { readUVarint } from "./varint.js";
+import { readHeader } from "./header.js";
 import {
   readInt8,
   readInt16,
@@ -111,23 +111,6 @@ export interface CompiledStream {
    * to the last complete row and returns what it has.
    */
   readRows: Reader<Row[]>;
-}
-
-/**
- * Read the `RowBinaryWithNamesAndTypes` header off the cursor: a LEB128 column
- * count, then that many column-name `String`s, then that many type-string
- * `String`s. Leaves the cursor at the first row's bytes.
- */
-export function readHeader(state: Cursor): {
-  names: string[];
-  types: string[];
-} {
-  const count = readUVarint(state);
-  const names: string[] = new Array(count);
-  for (let i = 0; i < count; i++) names[i] = readString(state);
-  const types: string[] = new Array(count);
-  for (let i = 0; i < count; i++) types[i] = readString(state);
-  return { names, types };
 }
 
 /**
