+++ skills/clickhouse-js-node-rowbinary-parser/src/reader.ts
@@ -49,4 +49,5 @@ export * from "./simpleAggregateFunction.js";
 export * from "./nested.js";
 export * from "./nothing.js";
 export * from "./aggregateFunction.js";
+export * from "./header.js";
 export * from "./compile.js";
